used maven-shade-plugin instead of maven-jar-plugin to create the executable jar file.
The old way involved referencing a bunch of jars in the maven repository via the Class-Path manifest.mf attribute.  However, the approach I was taking did not account for spaces in the paths, which is common on windows. (Spaces are used to separate classpath entries in manifest.mf.)

The new way doesn't reference any of the jars in the local maven repository; instead, we load all the dependencies up into the executable jar.  The jar file is fatter, but that doesn't really matter.
As a side benefit, one does not need to have maven installed nor the project built on his/her laptop in order to be able to run the controller client; he/she just needs to have the jar.

git-svn-id: http://arkham.ccci.org/svn/java/trunk/commons/maintenance-filter@9263 6185d61c-cf3b-044e-8ca2-f4aa9eb47c87